feat: Add vitest as a dependency and refactor random hex string generator
