+++ package.json
@@ -25,6 +25,7 @@
     "@types/react-dom": "^19.1.3",
     "@wxt-dev/module-react": "^1.1.3",
     "typescript": "^5.8.3",
+    "vitest": "^3.1.4",
     "wxt": "^0.20.6"
   }
 }
