+++ src/lib/generateRandomHexString.ts
